entrenamiento_niñeravirtual2/requirements.txt
dj-database-url>=2.1
Django>=4.2,<5.0
gunicorn>=21.2
numpy>=1.24
opencv-python>=4.8
pillow>=10.0
psycopg[binary]>=3.1 ; platform_system != "Windows"
python-dotenv>=1.0
requests>=2.31
torch>=2.1
ultralytics>=8.0
whitenoise>=6.6
